big update
rather than creating the archive on the server being backed up, we now upload the entire webroot to the sftp server and then execute the tar command on the sftp server. this helps if the server you're backing up has low disk space and can't fit two more copies of the files at once.